Add `Table` material and send `C2M_CROSS`
