--- Assets/_Image/_Phases/GameSettle.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 3c3e86e14aa2d486fb3fd5328eaf4944
-folderAsset: yes
-timeCreated: 1508572997
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
