--- Assets/_Image/_Phases/LinkDevice.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 668706de8a6fe4ee1a04d0d13e9199af
-folderAsset: yes
-timeCreated: 1508572980
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
