+++ Assets/_Material/Planks1c.mat.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 9be73c12eca0b4c73b189788b9809278
+timeCreated: 1509282563
+licenseType: Free
+NativeFormatImporter:
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
