+++ Assets/_Script/_Phases/GamePlay/Enter_Edge.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 17c1108130a8840b0a346469da4cafa5
+timeCreated: 1509279830
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
