+++ Assets/_Script/_Phases/GamePlay/GamePlay_Manager.cs
@@ -8,6 +8,10 @@ public class GamePlay_Manager : MonoBehaviour {
 
 	// when the phase begin
 	void OnEnable () {
+		
+	}
 
+	public void LeaveTable() {
+		networkController.SendToServer (new Pocket (gameController.version, C2M_Command.C2M_CROSS, new int[6]{1, 1, 1, 1, 1, 1}));
 	}
 }
