+++ Assets/_Script/_Phases/GamePlay/Swipe.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 248d493332994401eb98f01fd6ac45d6
+timeCreated: 1509279744
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
