--- Assets/_Script/_Phases/GamePlay/Swipe_Manager.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 9ba9d06c74b654b5a97d8089c3f822ee
-timeCreated: 1509272274
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
